requirements.txt
Flask
fairseq
zmq
SoundFile
packaging
ffmpeg-python
